requirements.txt
numpy
scipy
pillow
h5py
Flask==0.12
tensorflow
keras
matplotlib
